+++ app/controllers/app.js
@@ -77,7 +77,7 @@ var tableCache = LRU({
 var loggerOpts = {
     buffer: true,
     format: global.settings.log_format ||
-            ':req[X-Real-IP] :method :req[Host]:url :status :response-time ms -> :res[Content-Type]'
+            ':remote-addr :method :req[Host]:url :status :response-time ms -> :res[Content-Type]'
 };
 
 if ( global.log4js ) {
@@ -286,8 +286,6 @@ function handleQuery(req, res) {
             req.profiler.done('init');
         }
 
-        var affectedTables;
-
         // 1. Get database from redis via the username stored in the host header subdomain
         // 2. Run the request through OAuth to get R/W user id if signed
         // 3. Get the list of tables affected by the query
@@ -406,7 +404,6 @@ function handleQuery(req, res) {
                     };
                     tableCache.set(sql_md5, tableCacheItem);
                 }
-                affectedTables = tables;
 
                 if ( !authenticated && tableCacheItem ) {
                     var affected_tables = tableCacheItem.affected_tables;
@@ -479,7 +476,6 @@ function handleQuery(req, res) {
                   skipfields: skipfields,
                   sql: sql,
                   filename: filename,
-                  affectedTables: affectedTables,
                   bufferedRows: global.settings.bufferedRows,
                   callback: params.callback,
                   abortChecker: checkAborted